refactor: extract savings rate trend card
